Fixed compilation warnings in the roo_display code.

+++ test/compactor_test.cpp
@@ -53,7 +53,7 @@ class MockWriter {
 
  private:
   std::vector<Write> expected_;
-  int idx_;
+  size_t idx_;
 };
 
 TEST(Compactor, RightSimple) {
